Update SnakePanel.java
There are currently 5 levels. Right now you go to the next level when the snakes 'ate' 15 pieces. The length of the snakes determines the points and they get added up after each level and the snake with the highest score after level 5 wins. When a snake dies it just stops moving and so it can't get any more points this level, but it can play again in the next level. I'm not sure if this is how we wanted to do it so tell me if I should change something